+++ tests/unit/lib/Controller/SlaveControllerTest.php
@@ -23,6 +23,7 @@
 namespace OCA\GlobalSiteSelector\Tests\Unit\Controller;
 
 use Firebase\JWT\JWT;
+use OC\Authentication\TwoFactorAuth\Manager;
 use OCA\GlobalSiteSelector\AppInfo\Application;
 use OCA\GlobalSiteSelector\Controller\SlaveController;
 use OCA\GlobalSiteSelector\GlobalSiteSelector;
@@ -69,6 +70,9 @@ class SlaveControllerTest extends TestCase {
 	/** @var ISession | \PHPUnit_Framework_MockObject_MockObject */
 	private $session;
 
+	/** @var Manager | \PHPUnit_Framework_MockObject_MockObject */
+	private $twoFactorManager;
+
 	/** @var SlaveService | \PHPUnit_Framework_MockObject_MockObject */
 	private $slaveService;
 
@@ -88,6 +92,7 @@ class SlaveControllerTest extends TestCase {
 		$this->userBackend = $this->getMockBuilder(UserBackend::class)
 			->disableOriginalConstructor()->getMock();
 		$this->session = $this->createMock(ISession::class);
+		$this->twoFactorManager = $this->createMock(Manager::class);
 		$this->slaveService = $this->createMock(SlaveService::class);
 	}
 
@@ -104,6 +109,7 @@ class SlaveControllerTest extends TestCase {
 					$this->gss,
 					$this->userSession,
 					$this->session,
+					$this->twoFactorManager,
 					$this->urlGenerator,
 					$this->crypto,
 					$this->tokenHandler,